// .pnpm-store/v10/projects/82965466ecb203a6f7e1a64465070b57/src/features/settings/types.ts
export interface SiteSetting {
  id: number
  siteName: string
  siteSubtitle: string | null
  siteDescription: string | null
  siteKeywords: string | null
  logo: string | null
  favicon: string | null
  githubUrl: string | null
  aboutMarkdown: string | null
  icp: string | null
  publicEmail: string | null
}

export interface SiteSettingSubmitPayload {
  siteName: string
  siteSubtitle: string | null
  siteDescription: string | null
  siteKeywords: string | null
  logo: string | null
  favicon: string | null
  githubUrl: string | null
  aboutMarkdown: string | null
  icp: string | null
  publicEmail: string | null
}